chore: organized deepgram package into a folder
