+++ go.mod
@@ -8,3 +8,8 @@ require (
 )
 
 require github.com/Jeffail/gabs/v2 v2.6.1
+
+require (
+	github.com/google/go-cmp v0.5.9 // indirect
+	golang.org/x/xerrors v0.0.0-20220907171357-04be3eba64a2 // indirect
+)
